Move Android progress sections package
